Declined invitations and removed users now get removed from the UI without a reload
